SE3_EKF_Mango_Yield_Estimation/requirements.txt
numpy
opencv-python
opencv-contrib-python # For some tracking/stabilization algorithms if needed
scipy
pyyaml
torch # Or torch torchvision torchaudio specific to your system
torchvision
torchaudio
ultralytics # For YOLOv8, if used directly